Move statement related methods to a dedicated file
